Implemented the basics of the modifier geometry operation
Two properties where declared for the BillboardData structure:
a reference object and a billboard thickness parameter.
These two new properties have been added to the user interface.

With the addition of the reference object, two new functions
where added to the modifier declaration: updateDepgraph and
foreachObjectLink.
The first function allows the modifier to respond to changes to
the reference object in real time.
The second function seems to be related to loading and linking.

The MOD_mirror.c and blenkernel/BKE_modifier.h source files where
useful in determining how to add a reference object.